requirements.txt
streamlit
pandas
plotly
numpy
gspread
google-auth
google-auth-oauthlib
google-auth-httplib2
python-dateutil
